chore: Modularise Errors into PrimerFoundation
