+++ Packages/PrimerFoundation/Sources/PrimerFoundation/Extensions/Data.swift
@@ -1,18 +1,20 @@
 //
-//  DataExtension.swift
+//  Data.swift
 //
-//  Copyright © 2025 Primer API Ltd. All rights reserved. 
+//  Copyright © 2026 Primer API Ltd. All rights reserved. 
 //  Licensed under the MIT License. See LICENSE file in the project root for full license information.
 
 import Foundation
 
-internal extension Data {
+public extension Data {
     // NSString gives us a nice sanitized debugDescription
     var prettyPrintedJSONString: String? {
         guard let object = try? JSONSerialization.jsonObject(with: self, options: []),
               let data = try? JSONSerialization.data(withJSONObject: object, options: [.prettyPrinted, .sortedKeys]),
-              let prettyPrintedString = NSString(data: data,
-                                                 encoding: String.Encoding.utf8.rawValue)
+              let prettyPrintedString = NSString(
+                  data: data,
+                  encoding: String.Encoding.utf8.rawValue
+              )
         else { return nil }
 
         return prettyPrintedString as String
