+++ Packages/PrimerFoundation/Sources/PrimerFoundation/Model/PrimerSessionIntent.swift
@@ -1,7 +1,7 @@
 //
-//  PrimerFlowEnums.swift
+//  PrimerSessionIntent.swift
 //
-//  Copyright © 2025 Primer API Ltd. All rights reserved. 
+//  Copyright © 2026 Primer API Ltd. All rights reserved. 
 //  Licensed under the MIT License. See LICENSE file in the project root for full license information.
 
 public enum PrimerSessionIntent: String, Codable {
