+++ Sources/PrimerSDK/Classes/Error Handler/ErrorExtension.swift
@@ -5,28 +5,6 @@
 //  Licensed under the MIT License. See LICENSE file in the project root for full license information.
 
 import Foundation
-
-extension Array where Element == Error {
-
-    var combinedDescription: String {
-        var message: String = ""
-
-        self.forEach { err in
-            if let primerError = err as? (any PrimerErrorProtocol) {
-                message += "\(primerError.localizedDescription) | "
-            } else {
-                let nsErr = err as NSError
-                message += "Domain: \(nsErr.domain), Code: \(nsErr.code), Description: \(nsErr.localizedDescription) | "
-            }
-        }
-
-        if message.hasSuffix(" | ") {
-            message = String(message.dropLast(3))
-        }
-
-        return "[\(message)]"
-    }
-}
 import PrimerFoundation
 
 extension Error {
