+++ Sources/PrimerSDK/Classes/Extensions & Utilities/StringExtension.swift
@@ -5,10 +5,6 @@
 //  Licensed under the MIT License. See LICENSE file in the project root for full license information.
 
 import Foundation
-
-public extension String {
-    static var uuid: String {  UUID().uuidString }
-}
 import PrimerFoundation
 
 extension String {
